requirements.txt
# core
b2sdk==2.9.4
boto3==1.39.15
python-dotenv==1.1.1

# dev
ruff==0.12.5